+++ AppboyKit/headers/AppboyKitLibrary/ABKInAppMessageFull.h
@@ -1,0 +1,5 @@
+#import "ABKInAppMessageImmersive.h"
+
+@interface ABKInAppMessageFull : ABKInAppMessageImmersive
+
+@end