+++ AppboyKit/headers/AppboyKitLibrary/ABKInAppMessageFullViewController.h
@@ -1,0 +1,4 @@
+#import "ABKInAppMessageImmersiveViewController.h"
+
+@interface ABKInAppMessageFullViewController : ABKInAppMessageImmersiveViewController
+@end