+++ AppboyKit/headers/AppboyKitLibrary/ABKInAppMessageModal.h
@@ -1,0 +1,5 @@
+#import "ABKInAppMessageImmersive.h"
+
+@interface ABKInAppMessageModal : ABKInAppMessageImmersive
+
+@end