+++ AppboyKit/headers/AppboyKitLibrary/ABKInAppMessageModalViewController.h
@@ -1,0 +1,4 @@
+#import "ABKInAppMessageImmersiveViewController.h"
+
+@interface ABKInAppMessageModalViewController : ABKInAppMessageImmersiveViewController
+@end