+++ AppboyKit/headers/AppboyKitLibrary/ABKInAppMessageSlideupViewController.h
@@ -1,0 +1,9 @@
+#import "ABKInAppMessageViewController.h"
+
+@class ABKLabel;
+
+@interface ABKInAppMessageSlideupViewController : ABKInAppMessageViewController
+
+@property (retain, nonatomic) IBOutlet UIImageView *arrowImage;
+
+@end