+++ AppboyKit/headers/AppboyKitLibrary/ABKSlideupView.h
@@ -1,4 +1,0 @@
-#import <UIKit/UIKit.h>
-
-@interface ABKSlideupView : UIView
-@end